Begin implementatie SQL bestelling manager + aanpassen sql statements van bestelling en bestellingdetails

+++ KlantBestellingADO/SQL/dbo.bestelling.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[bestelling] (
     [bestellingID] INT             IDENTITY (1, 1) NOT NULL,
     [betaald]      BIT             NULL,
-    [prijsBetaald] DECIMAL (18, 2) NOT NULL,
+    [prijsBetaald] FLOAT           NOT NULL,
     [tijdstip]     DATETIME        NOT NULL,
     [klantID]      INT             NULL,
     PRIMARY KEY CLUSTERED ([bestellingID] ASC),
